+++ .env.example
@@ -2,3 +2,5 @@
 # Copy this file to .env.local and fill in your actual values
 NEXT_PUBLIC_SUPABASE_URL=your_supabase_project_url
 NEXT_PUBLIC_SUPABASE_ANON_KEY=your_supabase_anon_key
+
+NEXT_PUBLIC_API_URL=http://localhost:4000
